Refactor SvelteXmlBlocks
