--- src/test/resources/dev/blachut/svelte/lang/format/EnterInEmptyBlock.svelte
@@ -1,1 +1,0 @@
-{#if x}<caret>{/if}
