--- src/test/resources/dev/blachut/svelte/lang/format/EnterInEmptyBlock_after.svelte
@@ -1,3 +1,0 @@
-{#if x}
-    <caret>
-{/if}
